Pass after API changes (#224)
* examples: remove copied source, just leave links

Copied source code will always be out of sync and either wrong and/or
confusing. I therefore suggest to only link to the source files which
are up-to-date and tested as part of CI.

* guide: pass through docs after API changes

There have been quite a few API changes recently, so it was time to give
this a rough update.